requirements.txt
numpy~=2.2.4
opencv-python~=4.11.0.86
torchvision~=0.21.0
torch~=2.6.0
pillow~=11.1.0
transformers~=4.50.0
tqdm~=4.67.1
scikit-learn~=1.6.1
faiss-cpu~=1.10.0